Adding third pipeline for user service.
finl touch to manifest changing image name.

finl touch to manifest changing number of replicas.

changed the http port in service.

testing the 3 pipelines from zero to hero.

changed the name of the db in users.py.

changed the name of the db in the path users.py.

testing in paralel

fixing error in path to dockerfile in pipeline.

changing the name of the db.

uncomment few jobs in the test module.

uncomment few jobs in the test module.